- consolidating sensors migrations

+++ NEMO/apps/sensors/migrations/0001_initial.py
@@ -60,7 +60,7 @@ class Migration(migrations.Migration):
             fields=[
                 ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('created_date', models.DateTimeField(auto_now_add=True)),
-                ('value', models.BigIntegerField()),
+                ('value', models.FloatField()),
                 ('sensor', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, to='sensors.sensor')),
             ],
             options={
